Move `PathSource` to fixtures
